+++ lib/src/xml/nodes/node.dart
@@ -3,12 +3,11 @@ library xml.nodes.node;
 import '../mixins/has_attributes.dart';
 import '../mixins/has_children.dart';
 import '../mixins/has_parent.dart';
+import '../mixins/has_text.dart';
 import '../mixins/has_visitor.dart';
 import '../mixins/has_writer.dart';
-import '../navigation/descendants.dart';
+import '../mixins/has_xml.dart';
 import '../utils/node_type.dart';
-import 'cdata.dart';
-import 'text.dart';
 
 /// Immutable abstract XML node.
 abstract class XmlNode extends Object
@@ -16,14 +15,10 @@ abstract class XmlNode extends Object
         XmlParentBase,
         XmlAttributesBase,
         XmlChildrenBase,
+        XmlHasText,
         XmlHasVisitor,
-        XmlHasWriter {
+        XmlHasWriter,
+        XmlHasXml {
   /// Return the node type of this node.
   XmlNodeType get nodeType;
-
-  /// Return the text contents of this node and all its descendants.
-  String get text => descendants
-      .where((node) => node is XmlText || node is XmlCDATA)
-      .map((node) => node.text)
-      .join();
 }