+++ test/builder_test.dart
@@ -22,9 +22,9 @@ void main() {
         builder.element('price', nest: 29.99);
       });
     });
-    final xml = builder.build();
-    assertDocumentTreeInvariants(xml);
-    final actual = xml.toString();
+    final document = builder.buildDocument();
+    assertDocumentTreeInvariants(document);
+    final actual = document.toString();
     const expected = '<?xml version="1.0" encoding="UTF-8"?>'
         '<?xml-stylesheet href="/style.css" type="text/css" title="default stylesheet"?>'
         '<bookstore>'
@@ -50,7 +50,7 @@ void main() {
       builder.cdata('cdata');
       builder.text('textual');
     });
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected = '<?xml version="1.0"?>'
@@ -73,7 +73,7 @@ void main() {
           isSelfClosing: true, nest: '!');
       builder.element('self-closing-false', isSelfClosing: false);
     });
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected = '<element>'
@@ -87,7 +87,7 @@ void main() {
   test('nested string', () {
     final builder = XmlBuilder();
     builder.element('element', nest: 'string');
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected = '<element>string</element>';
@@ -100,7 +100,7 @@ void main() {
       'ri',
       ['n', 'g']
     ]);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected = '<element>string</element>';
@@ -110,7 +110,7 @@ void main() {
     final builder = XmlBuilder();
     final nested = XmlElement(XmlName('nested'));
     builder.element('element', nest: nested);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     expect(xml.children[0].children[0].toXmlString(), nested.toXmlString());
     expect(xml.children[0].children[0], isNot(same(nested)));
@@ -122,7 +122,7 @@ void main() {
     final builder = XmlBuilder();
     final nested = XmlElement(XmlName('nested'));
     builder.element('element', nest: [nested, nested]);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     expect(xml.children[0].children[0].toXmlString(), nested.toXmlString());
     expect(xml.children[0].children[0], isNot(same(nested)));
@@ -136,7 +136,7 @@ void main() {
     final builder = XmlBuilder();
     final nested = XmlText('text');
     builder.element('element', nest: nested);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     expect(xml.children[0].children[0].toXmlString(), nested.toXmlString());
     expect(xml.children[0].children[0], isNot(same(nested)));
@@ -148,7 +148,7 @@ void main() {
     final builder = XmlBuilder();
     final nested = XmlText('text');
     builder.element('element', nest: [nested, nested]);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     expect(xml.children[0].children[0].text, 'texttext');
     expect(xml.children[0].children[0], isNot(same(nested)));
@@ -160,7 +160,7 @@ void main() {
     final builder = XmlBuilder();
     final nested = XmlComment('abc');
     builder.element('element', nest: nested);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     expect(xml.children[0].children[0].toXmlString(), nested.toXmlString());
     expect(xml.children[0].children[0], isNot(same(nested)));
@@ -172,7 +172,7 @@ void main() {
     final builder = XmlBuilder();
     final nested = XmlAttribute(XmlName('foo'), 'bar');
     builder.element('element', nest: nested);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     expect(xml.children[0].attributes[0].toXmlString(), nested.toXmlString());
     expect(xml.children[0].attributes[0], isNot(same(nested)));
@@ -189,7 +189,7 @@ void main() {
     final builder = XmlBuilder();
     final nested = XmlDocumentFragment([XmlText('foo'), XmlComment('bar')]);
     builder.element('element', nest: nested);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     expect(xml.children[0].children[0].toXmlString(),
         nested.children[0].toXmlString());
@@ -212,7 +212,7 @@ void main() {
       builder.text('');
       builder.text('def');
     });
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected = '<text>abcdef</text>';
@@ -226,7 +226,7 @@ void main() {
       builder.attribute('lang', 'en', namespace: uri);
       builder.element('element', namespace: uri);
     }, namespace: uri);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected =
@@ -243,7 +243,7 @@ void main() {
       builder.attribute('lang', 'en', namespace: uri);
       builder.element('element', namespace: uri);
     }, namespace: uri);
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected =
@@ -265,7 +265,7 @@ void main() {
       expect(() => builder.namespace('http://2.foo.com/', 'xmlns'),
           throwsArgumentError);
     });
-    final actual = builder.build().toString();
+    final actual = builder.buildDocument().toString();
     const expected = '<element/>';
     expect(actual, expected);
   });
@@ -276,7 +276,7 @@ void main() {
       expect(() => builder.namespace('http://2.foo.com/', 'foo'),
           throwsArgumentError);
     }, namespace: 'http://1.foo.com/');
-    final actual = builder.build().toString();
+    final actual = builder.buildDocument().toString();
     const expected = '<foo:element xmlns:foo="http://1.foo.com/"/>';
     expect(actual, expected);
   });
@@ -285,7 +285,7 @@ void main() {
     builder.element('element', nest: () {
       builder.namespace('http://1.foo.com/', 'foo');
     });
-    final actual = builder.build().toString();
+    final actual = builder.buildDocument().toString();
     const expected = '<element xmlns:foo="http://1.foo.com/"/>';
     expect(actual, expected);
   });
@@ -294,7 +294,7 @@ void main() {
     builder.element('element', nest: () {
       builder.namespace('http://1.foo.com/', 'foo');
     });
-    final actual = builder.build().toString();
+    final actual = builder.buildDocument().toString();
     const expected = '<element/>';
     expect(actual, expected);
   });
@@ -310,7 +310,7 @@ void main() {
         });
       });
     });
-    final actual = builder.build().toString();
+    final actual = builder.buildDocument().toString();
     const expected = '<element xmlns:foo="http://1.foo.com/">'
         '<outer xmlns:foo="http://1.foo.com/">'
         '<inner xmlns:foo="http://1.foo.com/" foo:lang="en"/>'
@@ -330,7 +330,7 @@ void main() {
         });
       });
     });
-    final actual = builder.build().toString();
+    final actual = builder.buildDocument().toString();
     const expected = '<element xmlns:foo="http://1.foo.com/">'
         '<outer>'
         '<inner foo:lang="en"/>'
@@ -347,7 +347,7 @@ void main() {
         builder.element('inner', namespace: 'http://1.foo.com/');
       });
     });
-    final actual = builder.build().toString();
+    final actual = builder.buildDocument().toString();
     const expected = '<element xmlns:foo="http://1.foo.com/">'
         '<outer>'
         '<foo:inner/>'
@@ -358,7 +358,7 @@ void main() {
   test('entities cdata escape', () {
     final builder = XmlBuilder();
     builder.element('element', nest: '<test><![CDATA[string]]></test>');
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected =
@@ -370,7 +370,7 @@ void main() {
     builder.declaration(
         version: '0.5', encoding: 'ASCII', attributes: {'foo': 'bar'});
     builder.element('data');
-    final xml = builder.build();
+    final xml = builder.buildDocument();
     assertDocumentTreeInvariants(xml);
     final actual = xml.toString();
     const expected = '<?xml version="0.5" encoding="ASCII" foo="bar"?><data/>';
@@ -380,5 +380,22 @@ void main() {
     final builder = XmlBuilder();
     expect(() => builder.element('data', nest: builder.declaration),
         throwsA(isXmlNodeTypeException));
+  });
+  test('incomplete builder', () {
+    final builder = XmlBuilder();
+    builder.element('element', nest: () {
+      expect(builder.buildDocument, throwsStateError);
+    });
+    final document = builder.buildDocument();
+    expect(document.toString(), '<element/>');
+  });
+  test('resused builder', () {
+    final builder = XmlBuilder();
+    builder.element('element-one');
+    final firstDocument = builder.buildDocument();
+    expect(firstDocument.toString(), '<element-one/>');
+    builder.element('element-two');
+    final secondDocument = builder.buildDocument();
+    expect(secondDocument.toString(), '<element-two/>');
   });
 }