+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## 4.2.0
 
 * Deprecate standalone `XmlDocument parse(String input)` method, and introduce factory methods in the respective nodes `XmlDocument.parse(String input)` and `XmlDocumentFragment.parse(String input)`.
-* Deprecate `text` accessor on `XmlNode`; and introduce both getters and setters for `innerText`, `innerXml` and `outerXml`.
-* Add `replace(XmlNode other)` helper to make it easy to replace nodes in an existing tree.
-* Add `getElement(String name)` as a shortcut to find the first child element with a given name.
+* Introduce getters and setters for `XmlNode.innerText` (in most cases an alias to `XmlNode.text`), `XmlNode.innerXml` and `XmlNode.outerXml`.
+* Remove the `XmlDocument.text` override, which returned `null`.
+* Add `XmlNode.replace(XmlNode other)` to make it easier to replace nodes in an existing tree.
+* Add `XmlNode.getElement(String name)` as a shortcut to find the first child element with a given name.
 
 ## 4.1.0